requirements.txt
astroid==2.4.2
Django==2.1.7
djangorestframework==3.11.0
Faker==4.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.1
wrapt==1.12.1
